Add extras for cython
If you install the package with [cython] tag, this will enforce that
you're using a version of cython that actually compiles.  If you just
install without the cython tag but have an older version of cython in
your environment, the installation will fail but at least you'll have
some breadcrumbs in the setup.py to see what you ought to try (ie,
upgrade your cython).